Remove the `darc vmr update` command
