+++ src/Microsoft.DotNet.Darc/DarcLib/VirtualMonoRepo/VmrManagerBase.cs
@@ -1,7 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
@@ -9,8 +8,6 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.DarcLib.Helpers;
-using Microsoft.DotNet.DarcLib.Models;
-using Microsoft.DotNet.DarcLib.Models.Darc;
 using Microsoft.DotNet.DarcLib.Models.VirtualMonoRepo;
 using Microsoft.Extensions.Logging;
 
@@ -24,17 +21,13 @@ public abstract class VmrManagerBase
         "was interrupted. A new sync should start from {original} branch.";
 
     private readonly IVmrInfo _vmrInfo;
-    private readonly ISourceManifest _sourceManifest;
     private readonly IVmrDependencyTracker _dependencyInfo;
     private readonly IVmrPatchHandler _patchHandler;
-    private readonly IVersionDetailsParser _versionDetailsParser;
     private readonly IThirdPartyNoticesGenerator _thirdPartyNoticesGenerator;
     private readonly ICodeownersGenerator _codeownersGenerator;
     private readonly ICredScanSuppressionsGenerator _credScanSuppressionsGenerator;
     private readonly ILocalGitClient _localGitClient;
     private readonly ILocalGitRepoFactory _localGitRepoFactory;
-    private readonly IDependencyFileManager _dependencyFileManager;
-    private readonly IBasicBarClient _barClient;
     private readonly IFileSystem _fileSystem;
     private readonly ILogger _logger;
 
@@ -42,33 +35,25 @@ public abstract class VmrManagerBase
 
     protected VmrManagerBase(
         IVmrInfo vmrInfo,
-        ISourceManifest sourceManifest,
         IVmrDependencyTracker dependencyInfo,
         IVmrPatchHandler vmrPatchHandler,
-        IVersionDetailsParser versionDetailsParser,
         IThirdPartyNoticesGenerator thirdPartyNoticesGenerator,
         ICodeownersGenerator codeownersGenerator,
         ICredScanSuppressionsGenerator credScanSuppressionsGenerator,
         ILocalGitClient localGitClient,
         ILocalGitRepoFactory localGitRepoFactory,
-        IDependencyFileManager dependencyFileManager,
-        IBasicBarClient barClient,
         IFileSystem fileSystem,
         ILogger<VmrUpdater> logger)
     {
         _logger = logger;
         _vmrInfo = vmrInfo;
-        _sourceManifest = sourceManifest;
         _dependencyInfo = dependencyInfo;
         _patchHandler = vmrPatchHandler;
-        _versionDetailsParser = versionDetailsParser;
         _thirdPartyNoticesGenerator = thirdPartyNoticesGenerator;
         _codeownersGenerator = codeownersGenerator;
         _credScanSuppressionsGenerator = credScanSuppressionsGenerator;
         _localGitClient = localGitClient;
         _localGitRepoFactory = localGitRepoFactory;
-        _dependencyFileManager = dependencyFileManager;
-        _barClient = barClient;
         _fileSystem = fileSystem;
     }
 
@@ -187,149 +172,7 @@ public abstract class VmrManagerBase
         _logger.LogInformation("Committed in {duration} seconds", (int)watch.Elapsed.TotalSeconds);
     }
 
-    protected async Task CommitAmendAsync()
-    {
-        _logger.LogInformation("Amending the commit..");
-
-        var watch = Stopwatch.StartNew();
-
-        await _localGitClient.CommitAmendAsync(_vmrInfo.VmrPath);
-
-        _logger.LogInformation("Amended the commit in {duration} seconds", (int)watch.Elapsed.TotalSeconds);
-    }
-
-    /// <summary>
-    /// Recursively parses Version.Details.xml files of all repositories and returns the list of source build dependencies.
-    /// </summary>
-    protected async Task<IEnumerable<VmrDependencyUpdate>> GetAllDependenciesAsync(
-        VmrDependencyUpdate root,
-        IReadOnlyCollection<AdditionalRemote> additionalRemotes,
-        bool lookUpBuilds,
-        CancellationToken cancellationToken)
-    {
-        var transitiveDependencies = new Dictionary<SourceMapping, VmrDependencyUpdate>
-        {
-            { root.Mapping, root },
-        };
-
-        var reposToScan = new Queue<VmrDependencyUpdate>();
-        reposToScan.Enqueue(transitiveDependencies.Values.Single());
-
-        _logger.LogInformation("Finding transitive dependencies for {mapping}:{revision}..", root.Mapping.Name, root.TargetRevision);
-
-        while (reposToScan.TryDequeue(out var repo))
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            var remotes = additionalRemotes
-                .Where(r => r.Mapping == repo.Mapping.Name)
-                .Select(r => r.RemoteUri)
-                .Append(repo.RemoteUri)
-                .Prepend(repo.Mapping.DefaultRemote)
-                .Distinct()
-                .OrderRemotesByLocalPublicOther();
-
-            IEnumerable<DependencyDetail>? repoDependencies = null;
-            foreach (var remoteUri in remotes)
-            {
-                try
-                {
-                    repoDependencies = (await GetRepoDependenciesAsync(remoteUri, repo.TargetRevision))
-                        .Where(dep => dep.SourceBuild is not null);
-                    break;
-                }
-                catch
-                {
-                    _logger.LogDebug("Could not find {file} for {mapping}:{revision} in {remote}",
-                        VersionFiles.VersionDetailsXml,
-                        repo.Mapping.Name,
-                        repo.TargetRevision,
-                        remoteUri);
-                }
-            }
-
-            if (repoDependencies is null)
-            {
-                _logger.LogInformation(
-                    "Repository {repository} does not have {file} file, skipping dependency detection.",
-                    repo.Mapping.Name,
-                    VersionFiles.VersionDetailsXml);
-                continue;
-            }
-
-            foreach (var dependency in repoDependencies)
-            {
-                if (!_dependencyInfo.TryGetMapping(dependency.SourceBuild.RepoName, out var mapping))
-                {
-                    throw new InvalidOperationException(
-                        $"No source mapping named '{dependency.SourceBuild.RepoName}' found " +
-                        $"for a {VersionFiles.VersionDetailsXml} dependency of {dependency.Name}");
-                }
-
-                ProductConstructionService.Client.Models.Build? build = null;
-                if (lookUpBuilds)
-                {
-                    var builds = (await _barClient.GetBuildsAsync(dependency.RepoUri, dependency.Commit))
-                        .OrderByDescending(b => b.DateProduced)
-                        .ToList();
-
-                    if (builds.Count > 1)
-                    {
-                        _logger.LogInformation(
-                            "Found {number} builds for repo {repo} and commit {commit}. Will use the latest one.",
-                            builds.Count,
-                            dependency.RepoUri,
-                            dependency.Commit);
-                    }
-
-                    build = builds.FirstOrDefault();
-                }
-
-                var update = new VmrDependencyUpdate(
-                    mapping,
-                    dependency.RepoUri,
-                    dependency.Commit,
-                    repo.Mapping,
-                    build?.AzureDevOpsBuildNumber,
-                    build?.Id);
-
-                if (transitiveDependencies.TryAdd(mapping, update))
-                {
-                    _logger.LogDebug("Detected {parent}'s dependency {name} ({uri} / {sha})",
-                        repo.Mapping.Name,
-                        update.Mapping.Name,
-                        update.RemoteUri,
-                        update.TargetRevision);
-
-                    reposToScan.Enqueue(update);
-                }
-            }
-        }
-
-        _logger.LogInformation("Found {count} transitive dependencies for {mapping}:{revision}..",
-            transitiveDependencies.Count,
-            root.Mapping.Name,
-            root.TargetRevision);
-
-        return transitiveDependencies.Values;
-    }
-
-    private async Task<IEnumerable<DependencyDetail>> GetRepoDependenciesAsync(string remoteRepoUri, string commitSha)
-    {
-        // Check if we have the file locally
-        var localVersion = _sourceManifest.Repositories.FirstOrDefault(repo => repo.RemoteUri == remoteRepoUri);
-        if (localVersion?.CommitSha == commitSha)
-        {
-            var path = _vmrInfo.VmrPath / VmrInfo.SourcesDir / localVersion.Path / VersionFiles.VersionDetailsXml;
-            var content = await _fileSystem.ReadAllTextAsync(path);
-            return _versionDetailsParser.ParseVersionDetailsXml(content, includePinned: true).Dependencies;
-        }
-
-        VersionDetails versionDetails = await _dependencyFileManager.ParseVersionDetailsXmlAsync(remoteRepoUri, commitSha, includePinned: true);
-        return versionDetails.Dependencies;
-    }
-
-    protected async Task UpdateThirdPartyNoticesAsync(string templatePath, CancellationToken cancellationToken)
+    private async Task UpdateThirdPartyNoticesAsync(string templatePath, CancellationToken cancellationToken)
     {
         var isTpnUpdated = (await _localGitClient
             .GetStagedFiles(_vmrInfo.VmrPath))
@@ -353,7 +196,6 @@ public abstract class VmrManagerBase
     /// Takes a given commit message template and populates it with given values, URLs and others.
     /// </summary>
     /// <param name="template">Template into which the values are filled into</param>
-    /// <param name="mapping">Repository mapping</param>
     /// <param name="oldSha">SHA we are updating from</param>
     /// <param name="newSha">SHA we are updating to</param>
     /// <param name="additionalMessage">Additional message inserted in the commit body</param>
