+++ test/Microsoft.DotNet.DarcLib.Codeflow.Tests/VmrMultipleRemotesTests.cs
@@ -17,71 +17,6 @@ internal class VmrMultipleRemotesTests : VmrTestsBase
     private NativePath SecondDependencyPath => CurrentTestDirectory / (Constants.DependencyRepoName + "2");
 
     /// <summary>
-    /// The dependency tree of repos in this test looks like:
-    /// 
-    /// installer
-    ///   └── dependency
-    ///         
-    /// We will have two copies of "dependency" in folders dependency1 and dependency2
-    /// We will synchronize to the first one and then to the second one
-    /// </summary>
-    /// <returns></returns>
-    [Test]
-    public async Task SynchronizationBetweenDifferentRemotesTest()
-    {
-        var vmrSourcesDir = VmrPath / VmrInfo.SourcesDir;
-        var installerFilePath = vmrSourcesDir / Constants.InstallerRepoName / Constants.GetRepoFileName(Constants.InstallerRepoName);
-        var dependencyFilePath = vmrSourcesDir / Constants.DependencyRepoName / Constants.GetRepoFileName(Constants.DependencyRepoName);
-
-        var versionDetailsPath = InstallerRepoPath / VersionFiles.VersionDetailsXml;
-        var versionDetailsContent = await File.ReadAllTextAsync(versionDetailsPath);
-        versionDetailsContent = versionDetailsContent.Replace(
-            CurrentTestDirectory / Constants.DependencyRepoName,
-            CurrentTestDirectory / Constants.DependencyRepoName + "1");
-        await File.WriteAllTextAsync(versionDetailsPath, versionDetailsContent);
-        await GitOperations.CommitAll(InstallerRepoPath, "Point VersionDetails.xml to first location");
-
-        await InitializeRepoAtLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
-
-        var expectedFilesFromRepos = new List<NativePath>
-        {
-            installerFilePath,
-            dependencyFilePath,
-        };
-
-        var expectedFiles = GetExpectedFilesInVmr(
-            VmrPath,
-            [
-                Constants.InstallerRepoName,
-                Constants.DependencyRepoName,
-            ],
-            expectedFilesFromRepos);
-
-        CheckDirectoryContents(VmrPath, expectedFiles);
-
-        // Prepare a new commit in the second dependency repo
-        await File.WriteAllTextAsync(
-            SecondDependencyPath / Constants.GetRepoFileName(Constants.DependencyRepoName),
-            "New content only in the second folder now");
-        await GitOperations.CommitAll(SecondDependencyPath, "New commit in the second dependency repo");
-
-        // Point installer's VersionDetails.xml to this new repo
-        var oldSha = await GitOperations.GetRepoLastCommit(FirstDependencyPath);
-        var newSha = await GitOperations.GetRepoLastCommit(SecondDependencyPath);
-
-        versionDetailsContent = await File.ReadAllTextAsync(versionDetailsPath);
-        versionDetailsContent = versionDetailsContent.Replace(
-            CurrentTestDirectory / Constants.DependencyRepoName + "1",
-            CurrentTestDirectory / Constants.DependencyRepoName + "2");
-        versionDetailsContent = versionDetailsContent.Replace(oldSha, newSha);
-        await File.WriteAllTextAsync(versionDetailsPath, versionDetailsContent);
-        await GitOperations.CommitAll(InstallerRepoPath, "Point VersionDetails.xml to second location");
-        await UpdateRepoToLastCommit(Constants.InstallerRepoName, InstallerRepoPath);
-
-        CheckFileContents(dependencyFilePath, "New content only in the second folder now");
-    }
-
-    /// <summary>
     /// In this test:
     ///   - We will have two copies of a repo in folders dependency1 and dependency2
     ///   - We will create a commit only in the second folder
